requirements.txt
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
click>=8.0.0
plotly>=5.0.0
sentence-transformers>=2.2.0
numpy>=1.21.0
groq>=0.4.0
nltk>=3.6.0
wordcloud>=1.8.0
arabic-reshaper>=2.1.0
python-bidi>=0.4.2